% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Post_process.R
\name{arrival_tf}
\alias{arrival_tf}
\title{Transform raw pigeon arrival characther data into prediods.}
\usage{
arrival_tf(raw_arrival)
}
\arguments{
\item{raw_arrival}{scrapped arrival variable}
}
\description{
Taks raw characther arrival data, processes it and outputs a lubridate
peridod.
}
