% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_races.R
\name{assemble_tbl}
\alias{assemble_tbl}
\title{Assembre race tables from xml documents}
\usage{
assemble_tbl(races_xml, css_query_tbl)
}
\arguments{
\item{races_xml}{xml documents with race tables geneerated from
\code{\link{race_table_parse}}.}

\item{css_query_tbl}{individual css_query table used to access the
individual race}
}
\value{
\code{list} with two tables: \code{race_results}
and \code{race_info}
}
\description{
Assembre race tables from xml documents
}
