% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium_fun.R
\name{connect_remDr}
\alias{connect_remDr}
\title{Connect to remote driver}
\usage{
connect_remDr()
}
\value{
Object class remote driver.
}
\description{
This function assumes a docker container was proviously started using
\code{\link{start_chrome_remDr}}. It connects R to the docker container and
returns a remote driver object class which is used to interact with websites.
}
