% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{create_start_end}
\alias{create_start_end}
\title{Create index to iterate over css queries by groups of n queries.}
\usage{
create_start_end(seq_length, rows)
}
\arguments{
\item{seq_length}{integer specifying the size of the groups}

\item{rows}{integer with number of queries to be submitted.}
}
\value{
matrix with two columns with start and end end indexes.
}
\description{
Takes in a data.frame of css queries to be submitted using RSelenium and
outputs a matrix with start and end index used to further loop over.
}
