% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium_fun.R
\name{do_remDr}
\alias{do_remDr}
\title{Wrapper function to interact with remote driver}
\usage{
do_remDr(remDr, link, FUN, FUN_input)
}
\arguments{
\item{remDr}{remote driver connection object created with
\code{\link{connect_remDr}}.}

\item{link}{url for the website to intereact with.}

\item{FUN}{user-specified functon}

\item{FUN_input}{input that will go in \code{FUN}.}
}
\value{
Object generated by \code{FUN}.
}
\description{
Wrapper function to do operations with the remote driver. Allows the use of
user defined function for a specic link. The idea, is to make the process of
remotely interacting with websites, as flexible as possible so that I can
pass any function of interest to the scrape or do operations on the
link specified.
}
