% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Post_process.R
\name{error_capture}
\alias{error_capture}
\title{Check if saved tables contain errors}
\usage{
error_capture()
}
\value{
data frame withe the query index where the error occured. Data frame
with 0 rows if no errors where captured.
}
\description{
This function takes a series of rds files where we stored the scraped data
and checks wether any errors where captured by the the \code{purrr::safely}
function.
}
