% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_query_build.R
\name{extract_orgs}
\alias{extract_orgs}
\title{Extract queries for organizations}
\usage{
extract_orgs(years, remDr)
}
\arguments{
\item{years}{tibble or data.drame with year information and year css
queries generated by \code{\link{extract_years}}.}

\item{remDr}{Object class remote driver previosly connected using
\code{\link{connect_remDr}}.}
}
\value{
\code{tibble} with years, organization name and css queries for
both years and organizations.
}
\description{
Fuction that clicks on each year and extracts the queries to click on each
organization which has organized races.
}
