% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_races.R
\name{extract_race_html_options}
\alias{extract_race_html_options}
\title{Parse html source with race tables}
\usage{
extract_race_html_options(css_query_tbl, remDr)
}
\arguments{
\item{css_query_tbl}{table with year and organization css queries, generated
by the function \code{\link{extract_orgs}}.}

\item{remDr}{Object class remote driver previosly connected using
\code{\link{connect_remDr}}.}
}
\value{
xml_nodeset of parsed html options to be used to build queries.
}
\description{
Parse html source with race tables
}
