% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_query_build.R
\name{extract_years}
\alias{extract_years}
\title{Extract css queries for race years}
\usage{
extract_years(parsed_html)
}
\arguments{
\item{parsed_html}{pardsed html page using \code{\link{get_page_source}}.}
}
\value{
tibble with css queries and relevant info.
}
\description{
Given a parsed html, it will scrape and return a list of css queries to click
throught the \code{Years} field.
}
