% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium_fun.R
\name{get_page_source}
\alias{get_page_source}
\title{Get htlm page source}
\usage{
get_page_source(remDr, link)
}
\arguments{
\item{remDr}{remote driver connection}

\item{link}{link to the desired webpage}
}
\value{
parsed html file.
}
\description{
Function to instruct loaded remote driver to connect to url and download page
source. Note that indiviual download of source html wihthin each specific
query is handled diferently.
}
