% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Post_process.R
\name{names_to_ids}
\alias{names_to_ids}
\title{Replace real names by cryptographically generated random identifiers.}
\usage{
names_to_ids(competitor_names)
}
\arguments{
\item{competitor_names}{Vector of competitor names.}
}
\value{
tibble with column with old names and column with of competitor
cryptographically generated random identifiers.
}
\description{
Takes in a vector of real names and replaces it with a randomly generated
identifier.
}
