% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_scraping_funs.R
\name{pigeon_query_builder}
\alias{pigeon_query_builder}
\title{Build css query table for years and organizations}
\usage{
pigeon_query_builder(remDr)
}
\arguments{
\item{remDr}{Object class remote driver previosly connected using
\code{\link{connect_remDr}}.}
}
\value{
tibble with year and organization info and corresponding css queries.
}
\description{
Takes in a remote driver selenium driver and scrapes the APRU website to
build a query table used to scrape individual race information.
}
