% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_scraping_funs.R
\name{pigeon_scraper}
\alias{pigeon_scraper}
\title{Scraping function to extract data from the APRU National Database. It uses
functionality from the \code{furrr} package to run multiple queries in
parallel.}
\usage{
pigeon_scraper()
}
\description{
This is a do all function. It creates a query table, scrapes and processes
data from the APRU website in parallel. It save the output in
\code{data/raw_data} in single \code{.rds} files by group of queries.
}
