% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plan_future}
\alias{plan_future}
\title{Function to plan parallelization plan.}
\usage{
plan_future(workers)
}
\arguments{
\item{workers}{number of cores to be used in parallel.}
}
\description{
In order to parallize multiple scraping tasks over many different cores, we
need to set our parallization plan. By default, \code{pigeonscraper} uses
\code{plan(multiprocess)} and the user can set the number of cores to be used
by setting the parameter workers.
}
