% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_races.R
\name{pre_process_tbls}
\alias{pre_process_tbls}
\title{Table pre-process and aggregation by organization and by year}
\usage{
pre_process_tbls(tbls_list)
}
\arguments{
\item{tbls_list}{list of tables generated from \code{\link{assemble_tbl}}.}
}
\value{
list of tables by year and organization containing both iformation on
individual races and results for each pigeon entry in an individual race.
}
\description{
Table pre-process and aggregation by organization and by year
}
