% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puppeteer_functions.R
\name{pup_get_source}
\alias{pup_get_source}
\title{Get htlm page source using puppeteer API}
\usage{
pup_get_source(link)
}
\arguments{
\item{link}{link to website.}
}
\value{
parsed html source file.
}
\description{
Gets html source of a website.
}
