% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{race_info}
\alias{race_info}
\title{Pigeon race metadata}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 74612 rows and 20 columns.
}
\usage{
race_info
}
\description{
A dataset containing metadata from each race reported in the National
Database of the American Pigeon Racing Union.
}
\keyword{datasets}
