% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{race_results}
\alias{race_results}
\title{Pigeon race results data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2729115 rows and 19 columns.
}
\usage{
race_results
}
\description{
Data set with race data indexed by race_id. Contains only the top 20% of
birds.
}
\keyword{datasets}
