% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium_fun.R
\name{remDr_go_to_link}
\alias{remDr_go_to_link}
\title{Connect remote driver to specified url}
\usage{
remDr_go_to_link(remDr, link)
}
\arguments{
\item{remDr}{remote driver connection}

\item{link}{link to the desired webpage}
}
\value{
no return
}
\description{
Connect remote driver to specified url
}
