% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium_fun.R
\name{start_chrome_remDr}
\alias{start_chrome_remDr}
\title{Start remote driver with docker using the system command line.}
\usage{
start_chrome_remDr(kill = FALSE)
}
\arguments{
\item{kill}{Bolean indicating weather a runing docker container with selenium
should be killed.}
}
\description{
Start remote driver with docker using the system command line.
}
