% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Post_process.R
\name{state_abb_trans}
\alias{state_abb_trans}
\title{Transform state names into state abbreviations in a string.}
\usage{
state_abb_trans(raw_string)
}
\arguments{
\item{raw_string}{string to be parsed by the function.}
}
\value{
String with state names replaced by their abbreviatures.
}
\description{
This function takes a string or vector of strings, captures if a name of a
State is presente and changes it to its corresponding abbreviation.
}
